Create solution.py
O(nm) solution, where n and m are the dimensions of the matrix.

For each row/column pair element, if the row is in the upper triangular matrix and the value is not equal to that at the first row at the appropriate column, return false. Similarly, if the value is in the lower triangular matrix (not including the main diagonal) and the value is not equal to the first column with appropriate row, return false.

If the loop exits, return true.