Create solution.py
O(nm) time solution with O(nm) space (cannot do in-place transpose since the matrix may not be square - need to allocate space). If not including the result matrix, then O(1).

First, get the number of rows and columns in the initial matrix, and allocate memory for the transpose matrix by swapping the rows and columns.

Running through the rows and columns, assign the column and row to the row and column element from the original matrix.

Finally, return the transpose.