Create solution.py
O(n) time and O(1) space solution.

First, create a dummy node with the head as its next pointer, and set current pointer to the dummy node.

While the current pointer is not null, if the current pointer's next and  current pointer's next next value are not null, and the next value and next next value are equal, set the next next to a temporary pointer. While the temp pointer and its next pointer exist and the values are the same, advance the temporary pointer to its next pointer. Set the current pointer's next pointer to the temp pointer's next pointer.

Otherwise, advance the current pointer to its next pointer.

Finally, return the next pointer of the dummy (which will be the head).