Create solution.py
O(n) time and O(n) space solution using a counter hashmap and running through the list twice.

Create a dummy (sentinel) node with head as its next pointer. Set the current pointer to the dummy node and initialize the counter dictionary.

While current has a next pointer, increment the count in the dictionary and advance the pointer.

Set the current pointer to the dummy again.

While current has a next pointer that is not null, if the count in the dictionary is greater than or equal to 2 (has a duplicate), set current's next pointer to current's next next pointer (skip over the duplicate node). Otherwise, just advance the next pointer.

Finally, return the dummy's next pointer (the head).