Create solution_mod.py
O(1) time solution.

Knowing that the maximum possible power of 2 for the input n is 2^30,  (since 2^31 - 1 is the max value), return whether 2^30 mod n is 0.

This is done by first checking if n is greater than 0, and then bitshifting 1 by 30 positions (equivalent in binary to 2^30), then modding by n. Return whether this value is 0 or not.