Create solution.py
O(n) time and O(h = log(n)) space solution using recursion.

First, if the root is null return null. 

If the root value is greater than the high value, then trim the left subtree. Otherwise, if the root value is lower, then trim the right subtree.

Finally, assign the left subtree as the trimmed value of the left, and the right subtree as the trimmed value of the right side.

Finally, return the root.