Create solution.py
O(n) time and space solution using DFS recursion stack.

First, define the DFS function with arguments for the particular node, if it is left, and the depth.

If the node is null, return the depth. If it is a left branch node, then the depth is the max of the current depth, the DFS on the right subtree with it set to not the left, and the incremented depth, and the DFS on the left subtree with it being left set to true, and depth of 0. Otherwise, the depth is the max of the current depth and the DFS on the left subtree with it set to true and the depth incremented by 1, and the right subtree with it not being a left branch and depth reset to 0. Finally, the depth is returned.

Return the max of the DFS performed on the left subtree and the right subtree with depths set to 0.