Create solution.py
O(n) time and O(n) space solution using iterative BFS with queue.

First, if the root node is empty return an empty list.

Initialize a dictionary of column coordinates to the nodes, and the min and max values. Initialize the queue with a coordinate of 0 as the root node's x value, and the root node itself. Also, initialize the result list.

While the queue is not empty, get the x-coordinate and the node by popping. Append the node value to the appropriate x-coordinate key in the dictionary. Update the min and max x-value range.

If the left or right node exists, append x (-1 for left, +1 for right) and the node to the queue.

For each column (x-coordinate) in the range from the min to max x-values, append the list of values at the x-value to the result.

Return the result.