Create solution.py
O(n log n) average time solution, O(n) time worst-case due to linearity of duplicates. O(1) space.

First, initialize the left and right pointers to the end of the array. Also, set the result to the first value of the array.

While the pointers don't cross, calculate the midpoint. Update the output to the minimum of the output itself and the value at the midpoint.

If the midpoint is larger than the right pointer value, set the search area to the right side (update the left to mid + 1), else, if the midpoint value is less than the right value, set the search area to the left side (update right to mid - 1), otherwise, decrement the right pointer by 1 (this adds O(n) linearity for duplicates).

Finally, return the output.