Create solution.py
Create a function to rotate the matrix in-place in O(n^2) time and O(1) space.

The defined function goes through while the pointers don't cross, and rotates the matrix 90 degrees.

In the bottom loop, if the matrix is equal to the target, return true. Otherwise, rotate the matrix.

Return false if the loop exits.