Create solution.py
O(n log n) time and O(n) space using BFS with a queue. The O(n log n) comes from sorting, and it can be O(n log k), where k is the maximum width of the tree (which can at most be n/2 for a full binary tree at the lowest level (leaves).

First, set up a dictionary of column index to values and coordinates, the queue with initial values of (0, 0) for the row-column pair and the root node, and maximum and minimum values set respectively. Initialize the result array.

While the queue is not empty, get the row and column index pair and the node by popping.

Append to the particular column key the node value and row pair.

Adjust the max and min column ranges by updating them.

If the left or right child exists, append to the queue the incremented row (since the row is 1 level lower) and the column appropriately (-1 if to the left, +1 if to the right), along with the appropriate child.

Now, for each column in the range of min to max column index, get the value list from the column dictionary for that particular column. Now, sort the value by the row index first, then by the value itself. Append the list of values after this sorting to the result.

Finally, return the result.