Create solution.py
O(n) time and O(1) space solution using dynamically-sized sliding window.

Set the left and right pointers to 0, and the current sum and frequency of 1's to 0. Similarly, set the result to 0.

While the right pointer is not out of bounds, add the rightmost value to the current sum. If the rightmost value is a 1, reset the frequency of ones to 0. If the current sum is greater than the goal, remove the leftmost value and increment the left pointer (essentially decreasing the sliding window from the left). 

Now, while the pointers don't cross and the current sum is still equal to the goal, remove the leftmost element and increment the left pointer, and increment the frequency of ones.

Add this frequency of ones to the result, and then increment the right pointer.

Finally, return the result.