Create solution.py
O(n) time and O(k) space solution using a sliding window with a count dictionary, where n is the array size and k is the size of the input for the sliding window.

First, initialize the count dictionary and the running sum and the maximum sum values.

For the range of the first k values, add to the current sum and the count dictionary the counts of each value encountered.

If the size of the set is k (i.e., there are no duplicates), set the max sum to the current running sum.

Now, iterate from index k up to the end of the array. Subtract the value at the i-k index from the current running sum, and also decrement the count of this value from the dictionary. If the count becomes 0, delete this key from the dictionary. Now, add the value at index i and increment its count in the dictionary. If the dictionary becomes size k, then update the max sum with the current running sum.

Finally, return the result.