Create solution.py
O(n) time and O(h = log(n)) space due to DFS reverse in-order recursion.

First, set the current sum to 0.

Define the DFS for a node with a null node returning. Set the current sum variable as nonlocal since it is being referenced outside. Then, perform DFS on the right node. Using a temporary variable, get the value at the node. Update the current node's value to the current running sum so far. Then, add the temp variable value to the current running sum. Then perform DFS on the left.

Run DFS on the root, then return the root.