Create solution.py
O(n) time and O(h) space solution for BST using iterative DFS reverse in-order traversal with a stack.

First, initialize the stack and the running sum to 0. Set the node to the root.

While the stack is not empty or the node is not empty, while the node is not empty, append to the stack and keep advancing the node pointer to its right child. Then, set node to the popped value from the stack. The node value is updated by adding the current sum to that value, and the current sum is updated to the value of the node. Then, the node pointer is advanced to the left child.

Finally, return the root.