Create solution.py
O(n) time and O(h = log(n)) space from DFS recursion.

First, set the result list to the root value.

Define the DFS function. If the node is null, return 0. Then, get the left max value and right max value by running DFS on each subtree, and comparing it to 0 (to take care of negative cases). Then, set the result value to the max of the result value itself and adding the current node's value + left and right max values. This calculates the value from splitting at the given node. Finally, return the node value plus the max of either the left max or right max (this return value calculates the value of the subtree from the given node when the split does not occur).

Perform DFS on the root, then return the result.