Create solution.py
O(n) time and space solution using preorder DFS recursively and a comma-delimited string with the string of integers for each node values and 'n' for the null nodes.

For the serialize function, perform a preorder traversal and serialize into a string by taking null nodes as 'n' and going from node -> left -> right order (preorder). Separate these by commas (',').

For deserialize, perform DFS for each value in the string delimited by commas. Set the variable i at the 0 index. If the value is an 'n', increment the counter and return none. Otherwise, set the node to a new TreeNode with the integer value and increment the counter. Subsequently, perform DFS on both the left and right sides. Return the node. 
Finally, return the value from running the DFS.