Create solution.py
O(n) for startsWith() and search(), where n is the length of the word or the prefix, and O(n) for insert(), where n is the length of the word.

Create a TrieNode class for the trie node, which contains a dictionary of children with the nodes that they point to as the value, and the value of the character as the key, as well as an indicator variable indicating whether a given node is the end of a word.

Initialize the Trie class with a root node.

For the insert() method, set the current node to the root. For each character in the word, if the character is not in the children of the node, add a trie node for that particular character. Otherwise, keep advancing the pointer to the child of that particular node. 
Outside the loop, finally set the end of the word variable to true at the last node.

For the search() and startsWith() functions, iterate through each character in the word or prefix respectively, and return false if the character is not found. Set the current pointer to the children dictionary's key's value. The difference is that search() returns whether end of word is true after the loop, whereas startsWith() will automatically return true if it exits the loop without returning false.