Create solution.py
O(nk) time to build the trie of the words list, where n is the number of words and k is the average length of the word, and O(m^2) to go through the range of the length of the text with m being the length of the text. The memory is alphabet size times the key length times the number of keys in the trie, which is 26 * 1 * n, where n is the number of keys in the words list.

First, create a trie of the words in the words list and initialize the result array.

Then, for the left pointer in the range of the text, set the node to the trie. For the right pointer going from the left pointer to the end of the text, if the value at the right pointer is not in the node's children, break from the inner loop. Otherwise, set the node to the child whose value is the character found. If the node happens to be the end of the word, append the interval from left to right.

Finally, return the result.