Create solution.py
A trie solution that implements the addWord() method in O(n) time for the length of the wordand search() in O(nm) time, where n is the length of the word and m is the number of characters in the alphabet.

First, create a TrieNode class with a dictionary of children and a boolean indicating whether it is the end of the word.

Initialize the WordDictionary class with the root node. 

For the addWord(), have a classic Trie method of adding a word while the character in the word exists, and setting the boolean variable to true at the end of the word.

For the  search() method, define a DFS backtracking algorithm to go through the length of the word starting from index j to the end. For the case of the wildcard character '.', go through each of the children nodes and call DFS with the index incremented. If the children are not available, return false. Finally, if the loop finishes, return whether the end of word variable is true or not. Call this DFS function starting from the root with index 0, and return the value.