Create solution_reverse.py
O(n) time solution using reversal of array.

If the rightmost value is less than the leftmost value, reverse the array.

Then, check if each adjacent element for the left side is less than or equal to the right side. If this is not the case, return false.

If the loop finishes without exiting, return true.