Create solution_heap.py
O(n *m) time solution where n is the number of tasks, and m is the idle time (n in the argument). The space is also O(n) due to the max heap and the queue.

First, create a dictionary of the counts of each task.

Create a max heap, negating the value from the count values. Then, heapify it. 

Initialize the time variable to 0 and create a queue.

While the max heap or queue is not empty, increment the time variable. If the max heap is not empty, add 1 to the value popped from the max heap (this is decrementing the number of values technically, since this is a negative value). If the count is not 0, append the newly updated count and the time + idle time to the queue.  If the queue is not empty and the 2nd value at the front of the queue is equal to the current time (i.e., a machine is finished), push the count to the max heap by popping from the queue.

Finally, return the time.