Create solution.py
O(n) time greedy solution with O(n) space for the counters dictionary.

Get the counter dictionary to count each character occurrence in tasks. 

Then, get the maximum of the counts.

Now, add the number of values that have the maximum count.

Calculate the number of intervals by multiplying (n + 1) with the max count - 1 plus the max count, and return the maximum of this and the length of the tasks array.