Create solution.py
O(1) time for postTweet(), follow(), and unfollow() operations and O(k) for getNewsFeed() methods (there is a O(10 log k) algorithm for the k followers due to the max heap (which is mimicked as a min heap with negated values in Python), but this is still dominated by the heapify O(k) operation). There is also O(k) storage for the tweet maps for each user ID -> [count (timestamp, tweet ID], as well as a map of follower to followee IDs in a hashset.

The count (timestamp) is initialized to 0, and the dictionary for the tweet and follower maps are initialized.

For postTweet(), get the userId in the tweet map and assign the count and tweetId. Increment the count for the max heap (which is equivalent to decrementing in a min heap).

For the getNewsFeed() method, initialize the heap and result. Also, in the follow map add the own user ID. 
For each followee ID in the follow map for the user, if the followee ID is in the tweet map, set the index to the last element in the list of tweets for this particular followee ID. Get the count and tweet ID from the tweet map for this index (this is similar to the merge k sorted lists algorithm using a heap). Finally, add the count, tweet ID, followee ID, and the index decremented by 1 for the next iteration to the heap.
Heapify the heap.
Now, while the heap is not empty and the result length is less than 10, get the count, tweet ID, followee ID, and index by popping from the heap. Append the tweet Id to the result list. If the index is not negative (there are still elements available), get the count and tweet ID from the tweet map for this followee ID and index, and push this to the heap (along with decrementing the index). 
Finally, return the result.

for follow() and unfollow(), for the follower ID the followee ID is added to the set or removed. For unfollow(), the check is done to see if in the values set the followee ID exists prior to the operation.