Create solution_dfs.py
O(2^n) solution exploring all subsets using DFS. Space complexity is O(h = n) where h is the height of the recursion tree.

First, define the DFS function with inputs i and total, where i is the index of the nums array. If i is the length of the nums array, return the total since the last element has already been explored. Return the DFS solutions of including the current element at index i and not including the current element at index i, and add them together. This is done by incrementing the i and XOR-ing the total to the value at the current index if the element is included, and just passing the total if not.

Finally, return the DFS using input i = 0 and total = 0.