Create solution_bitops.py
Perform bit operations to find the result. O(n) time and O(1) space.

First, notice that there is 2^(n-1) number of cases for values which will XOR and have a 1 in each position except the leading 0's for a given number's binary representation. Then, counting the decimal value of each binary position, this can be calculated. 

Set the result to 0. Then, for each number in the array, logical OR with the result.

Multiply the result by 2^(n-1), where n is the length of the array, or equivalently, by bitshifting (n-1) to the left. Return this value.