Create solution.py
O(n 2^n) solution using backtracking with O(n) space for the depth of the recursion from the backtracking tree. There are n elements to go through, and 2^n subsets possible.

First, initialize the result array and the subset array.

Define the DFS function taking in the index i of the nums array. If i is out-of-bounds, append a copy of the subset (so it is not modified) to the result and return. Append the current index element from the array to the subset, then run DFS on the incremented index (this is running DFS including the current element at position i). Then, pop from the subset this very same value and run DFS on the incremented index (this is running DFS excluding the current element at position i). The order of this code matters.

Finally, run DFS starting at index 0 and return the result.