Create solution.py
O(2^t) time solution, where t is the target value, and O(t) space for the depth of the recursion tree during backtracking.

Initialize the result array.

Define the DFS function with inputs of the index, a current array, and the total so far. If the total equals the target, append to the result a copy of the current array (so it doesn't get modified) and return. If i is OOB (out-of-bounds) or the total is greater than the target, return. Do the two cases for including/not including the current element: append the current value and recurse on the DFS function by passing in i (the current index to be included), the current array, and updating the total with the current element's value (since the current element is still included). Next, do another iteration by popping from the current array and performing DFS with the incremented index (the next element), and the current array and the total (without adding the current element). This is the branch that does not use the current element in the combination sum.

Run DFS with input of index 0 to start from the beginning of the array, an empty current array, and total equal to 0.

Return the result.