Create solution.py
O(k n^k) time solution (loose bound), but more precisely O(k nC_k) = O(k n! / ((n-k)! k!)), which is just k times the combination. The space complexity is O(k n) at any time due to recursion stack for a given value.

First, initialize the result array.

Define the backtrack function with start value and the combination array. If the length of the combination array is k, append a copy of it to the result and return. For the values in the range from the start to n, append the value to the combination and recursively call backtrack with i + 1 and the combination array passed in. Then, pop the value from the combination array for the next iteration.

Run the backtrack function starting from 1 with an empty array for the combination array.

Return the result.