Create solution.py
O(sqrt(n) log(n)) time solution using division and square root repetitively, with space complexity being O(log n) from the recursion.

First, initialize the result array.

Define backtrack function with inputs number and the combination list. If the combination list is not empty, append the current number to the combination list and append this to the result. For the values from 2 to the integer square root of the number (inclusive range), if the number is not divisible then continue. If the combination list is not empty or the index number is greater than or equal to the combination list's last element, recursively call the backtrack function on the current number divided by i, and the combination list updated to include the current value i.

Run the backtrack function with inputs n and an empty list, then return the result.