Create solution.py
O(n * n!) upper bound, or O(n * nPk) = O(n * n!/(n - k)! where k is equal to n, so all possible permutations of n! time, with O(n) space for the recursion depth.

First, initialize the result array. 

If the length of the nums array is 1 return nums slice of all elements (this is a quicker way than using nums.copy() since copy is slow.

For all the elements in nums, get the value n by popping the first value from the array. Set the permutations by recursively calling on this modified nums array. For each permutation in perms, append the popped n value. Extend the result with the permutations list. Then, append the n back to the nums array.

Return the result.