Create solution_backtrack.py
O(n * n!) time solution with O(n) space from recursion depth.

Initialize the result array.

Define the backtrack function starting from a start index. If the start index is equal to the size of the array, append the nums array to the result and return. For i in the range of the start index to the end of the array, swap the value at the start index and the current index. Then run backtrack on the incremented index. Swap back the values.

Run backtrack with start value of 0. Return the result.