Create solution.py
O(n log n) + O(n 2^n) = O(n 2^n) dominates for the time complexity, with O(n) recursion depth for the space complexity using backtracking.

First, initialize the result array, and sort the nums array.

Define the backtrack function with input i for the index in the nums array and the subset built so far. If i equals the size of the nums array, append the copy of the current subset to the result and return. Otherwise, append the current number at index i to the subset, then backtrack on this subset with index i incremented (this is including the current number in the future subsets). Pop the value from the subset, then while the index is not out-of-bounds and the number at index i equals the next number at index (i+1), keep incrementing the index (to look at non-duplicate subsets). Then, run backtrack on the incremented index and the subset (with the value popped).

Run backtrack starting at index 0 with an empty array as the initial subset input, then return the result.