Create solution.py
O(2^n) + O(n log n) = O(2^n) time and O(n) space solution using backtracking recursion.

First, initialize the result array.

Then, sort the candidates in order to make sure duplicate branches are not explored further.

Define the backtrack function with the start position of the backtrack, the current array of the values so far, and the target sum. If the target is 0, append the copy of the current array to the result and return if the target is less than 0. Set the previous value to -1 since all values are positive. For index i iterating through the range from the start position to the end of the array, if the array equals the previous value, go to the next iteration. Append to the current array the current element otherwise, and backtrack with the incremented index, the current array, and the target adjusted by subtracting the current (included) number. Then, pop the value that was added, and set the previous value to the current element value.

Run backtrack starting from position 0 index, an empty array, and the target sum.

Return the result.