Create solution.py
O(n!) time solution with O(n) recursion depth from backtracking DFS.

First, set the result array. Create a counter dictionary of the nums array. Initialize the permutation array.

Define the backtrack DFS function. If the length of the permutation equals the length of the nums array, append to the result a copy of the permutation array and return. For each number in the counter dictionary, if the count is greater than 0, append it to the permutation array and decrement from the count dictionary. Run DFS recursively on this new change (this is including that element in the permutation iteration). Then, increment the count in the count dictionary again and pop from the permutation (this is going back to running permutation in the other order).

Run the DFS, then return the result.