Create solution.py
O(2^n) solution where n is maximum 60. The k is also 2 through 9, so technically is constant time. The O(n) depth is the space complexity with n a maximum of 60, so also constant.

First, set the result array and the start and end points of the array to 1 through 9.

Define the backtrack function with input i as the initial position, n the number target, and the current array. If the length of the current array is k and the target n is 0, append a copy of the current array to the result and return. For indicies j in the range of the beginning position i to the end point 9 (inclusive), append the value to the current array. Run backtrack on the next number j + 1 with target adjusted by this value and the current array (this includes the current value). Then, pop the value from the current array for the next iteration (to not include the current value).

Run backtrack from the start position of 1, n as the target, and the empty array.

Return the result.