Create solution.py
O(2^n) time solution with O(n) space from the recursion using backtracking.

First, set the result array.

Define the backtrack function with the current index as input and the current array of characters so far. If the index i is the length of the input string, join the copy of the character array into a string and append to the result and return. Append the current character at index i to the current array and run backtrack with the incremented index. This is including the current character with its original case (if it is an alphabet character) or just the digit itself. Then, if the current character is an alphabet character, append the swapped case to the current array and run the backtrack recursion with the incremented index. Next, pop the value from the current array.

Run backtrack starting from position 0 with an empty array.

Return the result.