Create solution.py
O(n * 2^n) time complexity with O(n) space from backtrack recursive solution.

First, set the result array.

Define the backtrack function with the start position i, current array of strings, and the count. If i is equal to the length of the word, if the count is not 0, append this to the current array and append the joined string of the copy of the current array to the result and then pop the string. Otherwise, append the joined string of the copy of the current array. Return afterwards. If the i is not equal to the length of the word, backtrack on the incremented position with the current array, and the count incremented by 1 as well. Assign a string applied to the string count concatenated to the word's current i position if the count is not 0, otherwise just the character at position i. Append this to the current array and run backtrack with the incremented index i with the current array, but the count reset to 0. Then, pop this string from the current array.

Run backtrack with start index 0 and current array of 0, and a count of 0.

Return the result.