Create solution.py
O(n 2^n) time solution due to all combinations of 2^n strings explored n times each, with O(n) space from the visited set and recursive stack from backtracking DFS.

First, set the result to 0 and n to the length of the string. Also, create a visited set of visited substrings.

Define the DFS with input i as the start index. Specify the result is a nonlocal variable in order to access it within the function. If the length of the visited set + n - i is less than or equal to the result, the result cannot be improved, so return. If i is equal to n, set the result to the length of substrings seen so far and return.
For j in the range from the start position i to n + length of the visited set - result, get array of characters from start position i to the minimum of j and n-1 (inclusive), and join together as a string. If this substring is not visited yet, add it to the set, and perform DFS with incremented index j+1. Then remove the substring from the visited set.

Perform DFS starting with initial position 0, then return the result.