Create solution.py
O(n * m * 4^w) time complexity with n x m matrix board, and w the length of the word (with 4 directions). The space complexity is O(w) for both the recursive backtracking DFS call stack and the path set.

First, get the row and column sizes and initialize the path set.

Define DFS on a given row r, column c, with position i for the word index. If i equals the length of the word, return true. If r or c is out-of-bounds, or if the current character of the word at index i is not equal to the current character on the board at (r, c), or if (r, c) has already been visited (in the path set), return false. Otherwise, add (r, c) to the path set and perform DFS with incremented index i on all 4 adjacent directions (r +/- 1 and c +/- 1), and logical OR each of them and set it to the result. Finally, remove from the path set the coordinates just explored (r, c), and return the result.

Now, for each row/column pair on the board, if the DFS run at that coordinate starting with index 0 for the word is true, return true.

Otherwise, if the loop exits without returning, return false.