Create solution.py
O(2^n) time solution from partitioning each possible combination and O(n) space for the recursive backtracking DFS.

First, initialize the result array and the partition array.

Define isPalindrome function to check if given the left and right pointers, the string is a palindrome.

Define the DFS function with input index i at which to start. If i equals the length of the string, append the partition copy to the result and return. For j in the range of start index i to the end of the string, if the string from i to j is a palindrome, append the string partition i to j to the partition list. Perform DFS on the incremented index. Then pop from the partition for the next iteration.

Perform DFS starting from index 0, then return the result.