Create solution.py
O(n) time solution for each number in the array and O(1) space using bit manipulation.

First, initialize the toggle for bits that occur odd number of times and bits that occur even number of times as 0.

For each number in the array, XOR the odd bits with the number ANDed with the negation of the even bits. This makes it so that bits that appeared only once remain Similarly, XOR the even bits with the number ANDed by the negation of the odd bits. This makes it so that the value in evens keeps the bits that have appeared even number of times (twice).

Afterwards, the odd bits will contain the number that appears once - return this value.