Create solution.py
O(3^n = 3^4) time solution with O(n) = O(5) since the maximum height of the backtrack DFS tree is 5 (at most 4 groups of digits in an IP address). If the integer string is greater than 12 (since there is at most 12 digits possible), it is not a valid IP.

First, initialize the result array.

If the length of the string is greater than 12, return the result (empty array).

Define the backtrack function with start index i, the number of dots, and the current IP address string so far as parameters. If the number of dots is 4 and the index equals the length of the string, append the current IP address up to (but not including) the last character (which will be a '.') to the result and return. If the number of dots is greater than 4, this is not valid, so return. For index j in the range from the start position i to the minimum of i + 3 (since there is at most 3 digits) or the length of the string, if the integer form of the string from i to j inclusive is less than 256 and either: i equals j (this means it is a single digit) or if the first letter at index i is not a leading '0' character, then backtrack recurse on the incremented index, number of dots incremented, and the current IP with the string from i to j (inclusive) appended to it.

Run the backtrack function with start position 0, number of dots 0, and an empty string.

Return the result.