Create solution.py
O(n) time and O(1) space solution using bit manipulation.

First, get the XOR of all the values in the array. This will yield the XOR of the 2 numbers which are not in pairs. The differing bits will show that one number has a 1 at a given bit position while the other does not.

Now, initialize the differing bit to start at the LSB position.

While the XOR and the differing bit are not in the same position, left shift the differing bit by 1 to the position where there is a differing bit.

Initialize the differing bit matching and not matching values.

Now, running through the array again, if the differing bit position's 1 matches with the number, XOR this number with the differing match. Otherwise, XOR the number with the non-differing match.

Return the differing match (which will be one of the numbers whose first differing bit was found) and the non-differing match (which did not contain the first non-differing bit).