Create solution_cnt_intersection.py
O(nm) time solution where n is the number of words and m is the average length of a word/maximum length of the word. Space complexity is O(26), which is O(1) since there are only 26 lowercase alphabets.

First, set the result array and create a counter dictionary of the first word in the list.

For words in the rest of the list, create a counter of that word counting the number of occurrences of each character. For each character in the original count dictionary, find the minimum count of the original count dictionary and the current word's count dictionary. This acts like a defaultdict and yields 0 if the character does not exist. Therefore, this will act like an intersection of the characters and their counts.

Now, for each character remaining in the count dictionary, append to the list the number of times in the dictionary.

Finally, return the result.