Create solution_comb.py
O(n) time and space solution using a counter hashmap with combinatorics formula for counting the number of pairs.

First, initialize the result.

Create a counter dictionary to count the occurrence of each number.

For the values in the dictionary, apply the combination formula of nC2 and add this to the result. This works since the number of pairs that can be formed for each number given its occurrence count is the number of is n choose 2.

Return the result.