Create solution.py
O(n) time and space solution using a dictionary.

First, initialize the result to 0 and the empty count dictionary.

For each number in the array, if that number is already in the dictionary, add the count in the dictionary to the result and increment the count in the dictionary. Otherwise, set the key as the number and the value to 1.

Finally, return the result.